+++ publish/CHANGELOG.md
@@ -1,5 +1,9 @@
+- v1.2
+  - Fixed for the new update.
+  - Changed dungeon recovery to not remove any rooms.
+
 - v1.1
-	- Adds a fix for overfilled crafting structures.
+  - Adds a fix for overfilled crafting structures.
 
 - v1.0
-	- Initial release.
+  - Initial release.